+++ app/src/main/assets/CHANGELOG.md
@@ -6,6 +6,15 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Ypupiara](https://en.wikipedia.org/wiki/Ypupiara)
 
+### Next version
+
+* Added mini audio player
+* Added video preview image
+* Small improvements and minor bug fixes
+* Updated [AndroidX](https://developer.android.com/jetpack/androidx/versions/all-channel)
+* Updated [Public Suffix List](https://github.com/publicsuffix/list)
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
 ### 1.2151 - 2024-01-22
 
 * Fixed crash on some devices